Extrapolate as opposed to zero-pad cycle subseries #9
This isn't a complete solution, but gets us closer to being able to return
decent output for the degenerate case described in this issue. The root cause
was that we were just padding the cycle subseries with 0 on each side, when
according to the fortran, we should be doing single point Loess interpolation.
Due to limitations in the commons-math library, this uses another interpolator
which does not throw `OutOfRangeException` for datapoints outside the initial
time series range.

With 1 inner loop pass, the data still looks a little weird. However, with the
default 10, it looks pretty decent. So it seems like this is an improvement.