Use LoessInterpolator not NevilleInterpolator for padEdges #9
Also removed a line of code that was padding the cycle subseries. Not sure why
that was there, but I don't think that we need to artifically extend the cycle
subseries to handle edge cases, but instead handle them directly when we pad
the edges. The LoessInterpolator and dependent classes cannot work on something
with less than 2 points, and seems to have the tendency to return NaN for
coefficients of its polynomials. It seems like a problem with the library;
however, we can guard against it and degrade gracefully.